Show a more accurate message. Fix #147
